+++ src/models.py
@@ -1,4 +1,4 @@
-import random, randomname
+import random, randomname, game
 
 def _to_key(teamName):
 		key = teamName.lower().replace(" ","")
@@ -137,6 +137,43 @@ class DraftClass(object):
 
 	def get_draft_results(self):
 		return self.members
+
+class Season(object):
+
+	def __init__(self, league, schedule, week):
+		self.league = league
+		self.schedule = schedule
+		self.week = week
+		self.finished_games = []
+		self.standings = {}
+		self._init_standings()
+
+	def get_week(self):
+		return self.week
+
+	def get_remaining_weeks(self):
+		return (len(self.schedule) - self.week)
+
+	def _init_standings(self):
+		for team in self.league.get_all_teams():
+			self.standings[team.get_team_name()] = [0,0]
+
+	def get_standings(self):
+		return self.standings
+
+	def set_standings(self,value):
+		self.standings = value
+
+	def advance_week(self):
+		if(self.get_remaining_weeks() <= 0):
+			return
+		for match in self.schedule[self.week]:
+			away_team = self.league.get_team(match[0])
+			home_team = self.league.get_team(match[1])
+			result = game.play_game(away_team, home_team)
+			self.standings[result.get_winner().get_team_name()][0] += 1
+			self.standings[result.get_loser().get_team_name()][1] += 1
+		self.week += 1
