+++ src/rostertool.py
@@ -1,9 +1,10 @@
-from models import Player
-from models import Team
-from models import League
-import json, jsontool, randomname, random
+from models import Player, Team, League, Season
+import json, jsontool
 
 LEAGUE_DATA_PATH = "/Users/johnshea/Repos/nhasim_python/data/"
+
+def _to_file_name(name):
+	return name.replace(" ","")+".json"
 
 def write_league(league):
 	data = jsontool.league_to_json(league)
@@ -17,5 +18,9 @@ def load_league(leagueName):
 		data = json.load(datafile)
 		return jsontool.json_to_league(data)
 
-def _to_file_name(name):
-	return name.replace(" ","")+".json"
+def write_season(season):
+	data = jsontool.season_to_json(season)
+	filename = _to_file_name(season.league.get_league_name()+"Season")
+	with open(LEAGUE_DATA_PATH+filename,'w') as datafile:
+		json.dump(data,datafile)
+
