+++ src/seasonsim.py
@@ -1,7 +1,7 @@
 import game, rostertool
 from os import system
 from tabulate import tabulate
-
+from models import Season
 
 def _get_input():
 	inp = raw_input("\n>")
@@ -13,43 +13,6 @@ def _get_input():
 		cmd[i] = cmd[i].rstrip().lstrip()
 		cmd[i] = cmd[i].replace("_", " ")
 	return cmd
-
-class Season(object):
-
-	def __init__(self, league, schedule, week):
-		self.league = league
-		self.schedule = schedule
-		self.week = week
-		self.finished_games = []
-		self.standings = {}
-		self._init_standings()
-
-	def get_week(self):
-		return self.week
-
-	def get_remaining_weeks(self):
-		return (len(self.schedule) - self.week)
-
-	def _init_standings(self):
-		for team in self.league.get_all_teams():
-			self.standings[team.get_team_name()] = [0,0]
-
-	def get_standings(self):
-		return self.standings
-
-	def set_standings(self,value):
-		self.standings = value
-
-	def advance_week(self):
-		if(self.get_remaining_weeks() <= 0):
-			return
-		for match in self.schedule[self.week]:
-			away_team = self.league.get_team(match[0])
-			home_team = self.league.get_team(match[1])
-			result = game.play_game(away_team, home_team)
-			self.standings[result.get_winner().get_team_name()][0] += 1
-			self.standings[result.get_loser().get_team_name()][1] += 1
-		self.week += 1
 
 class SeasonShell(object):
 
@@ -72,12 +35,26 @@ class SeasonShell(object):
 		system("clear")
 		print_standings(self.season.get_standings())
 
+	def _help(self):
+		table = []
+		table.append(["advweek X","Advance X weeks ahead"])
+		table.append(["remweek","Get the number of weeks left"])
+		table.append(["quit","Quit the program"])
+		print(tabulate(table))
+
 	def run(self):
 		self._refresh()
 		while(True):
 			cmd = _get_input()
 			if(cmd[0] == "advweek"):
-				self._advance_week(int(cmd[1]))
+				if(len(cmd) > 1):
+					self._advance_week(int(cmd[1]))
+				else:
+					self._advance_week(1)
+			elif(cmd[0] == "remweek"):
+				print(self.season.get_remaining_weeks())
+			elif(cmd[0] == "help"):
+				self._help()
 			elif(cmd[0] == "quit"):
 				break
 			else:
@@ -90,7 +67,6 @@ class SeasonShell(object):
 		for i in range(amount):
 			self.season.advance_week()
 		self._refresh()
-
